analyzer: avoid global bailout on `typeof require`
Some modules use `typeof require === 'function'` for feature detection.
This would cause a global bailout because common-shake saw a `require`
in an unknown situation.
This case doesn't have any risk of including a module that we can't
statically determine in the bundle, so we can keep going.

To do this the walk() function now adds a `.parentNode` property to the
nodes as it walks the AST.